+++ src/webapp/package.json
@@ -17,6 +17,7 @@
     "react-bootstrap": "^2.0.0-beta.5",
     "react-dom": "^17.0.2",
     "react-dropzone": "^11.3.4",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "typescript": "^4.3.5",
     "web-vitals": "^1.1.2"
@@ -47,6 +48,7 @@
   },
   "devDependencies": {
     "@types/papaparse": "^5.2.6",
-    "@types/pdfmake": "^0.1.18"
+    "@types/pdfmake": "^0.1.18",
+    "@types/react-router-dom": "^5.3.1"
   }
 }
