+++ src/webapp/src/components/adminentries/RegistrationClassCount.tsx
