+++ src/webapp/src/components/results/Results.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { PageTitle } from '../PageTitle';
+
+export class Results extends React.Component<{}, {}, {}> {
+
+    render () {
+    return (
+        <div>
+            <PageTitle 
+                title="Results" 
+            />
+            <p>
+                A table of event results goes here. Or a list of classes if only one event.
+            </p>
+        </div>
+    );
+    }
+}
